fix code

+++ cyy_naive_pytorch_lib/pipelines/text_field.py
@@ -9,5 +9,6 @@ from torchtext.vocab import Vocab
 
 def get_text_and_label_fields():
     text_field = data.Field(tokenize="spacy", tokenizer_language="en_core_web_sm")
-    label_field = data.LabelField(dtype=torch.float)
+    # label_field = data.LabelField(dtype=torch.float)
+    label_field = data.LabelField()
     return (text_field, label_field)
